refactor: combine event and event attendee modules
