+++ src/database/data-source.ts
@@ -2,25 +2,25 @@ import * as dotenv from 'dotenv';
 import { DataSource } from 'typeorm';
 import { Canary } from '../canaries/models/canary.model';
 import { Comment } from '../comments/models/comment.model';
-import { EventAttendee } from '../events/event-attendees/models/event-attendee.model';
+import { EventAttendee } from '../events/models/event-attendee.model';
 import { Event } from '../events/models/event.model';
-import { GroupConfig } from '../groups/models/group-config.model';
-import { GroupMemberRequest } from '../groups/models/group-member-request.model';
 import { GroupRolePermission } from '../groups/group-roles/models/group-role-permission.model';
 import { GroupRole } from '../groups/group-roles/models/group-role.model';
+import { GroupConfig } from '../groups/models/group-config.model';
+import { GroupMemberRequest } from '../groups/models/group-member-request.model';
 import { Group } from '../groups/models/group.model';
 import { Image } from '../images/models/image.model';
 import { Like } from '../likes/models/like.model';
 import { Post } from '../posts/models/post.model';
 import { ProposalConfig } from '../proposals/models/proposal-config.model';
 import { Proposal } from '../proposals/models/proposal.model';
-import { ProposalAction } from '../proposals/proposal-actions/models/proposal-action.model';
 import { ProposalActionEventHost } from '../proposals/proposal-actions/models/proposal-action-event-host.model';
 import { ProposalActionEvent } from '../proposals/proposal-actions/models/proposal-action-event.model';
 import { ProposalActionGroupConfig } from '../proposals/proposal-actions/models/proposal-action-group-config.model';
 import { ProposalActionPermission } from '../proposals/proposal-actions/models/proposal-action-permission.model';
 import { ProposalActionRoleMember } from '../proposals/proposal-actions/models/proposal-action-role-member.model';
 import { ProposalActionRole } from '../proposals/proposal-actions/models/proposal-action-role.model';
+import { ProposalAction } from '../proposals/proposal-actions/models/proposal-action.model';
 import { ServerConfig } from '../server-configs/models/server-configs.model';
 import { ServerInvite } from '../server-invites/models/server-invite.model';
 import { ServerRolePermission } from '../server-roles/models/server-role-permission.model';
