+++ src/events/events.module.ts
@@ -2,17 +2,15 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Image } from '../images/models/image.model';
 import { Post } from '../posts/models/post.model';
-import { EventAttendeesModule } from './event-attendees/event-attendees.module';
-import { EventsResolver } from './events.resolver';
 import { EventsService } from './events.service';
+import { EventAttendee } from './models/event-attendee.model';
 import { Event } from './models/event.model';
+import { EventAttendeesResolver } from './resolvers/event-attendees.resolver';
+import { EventsResolver } from './resolvers/events.resolver';
 
 @Module({
-  imports: [
-    TypeOrmModule.forFeature([Event, Image, Post]),
-    EventAttendeesModule,
-  ],
-  providers: [EventsService, EventsResolver],
+  imports: [TypeOrmModule.forFeature([Event, EventAttendee, Image, Post])],
+  providers: [EventsService, EventsResolver, EventAttendeesResolver],
   exports: [EventsService, TypeOrmModule],
 })
 export class EventsModule {}
