+++ src/proposals/proposal-actions/models/proposal-action-event-host.model.ts
@@ -7,7 +7,7 @@ import {
   PrimaryGeneratedColumn,
   UpdateDateColumn,
 } from 'typeorm';
-import { EventAttendeeStatus } from '../../../events/event-attendees/models/event-attendee.model';
+import { EventAttendeeStatus } from '../../../events/models/event-attendee.model';
 import { User } from '../../../users/models/user.model';
 import { ProposalActionEvent } from './proposal-action-event.model';
