+++ src/proposals/proposal-actions/proposal-actions.module.ts
@@ -1,6 +1,6 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { EventAttendee } from '../../events/event-attendees/models/event-attendee.model';
+import { EventAttendee } from '../../events/models/event-attendee.model';
 import { Event } from '../../events/models/event.model';
 import { GroupRolesModule } from '../../groups/group-roles/group-roles.module';
 import { GroupsModule } from '../../groups/groups.module';
@@ -13,12 +13,12 @@ import { ProposalActionPermission } from './models/proposal-action-permission.mo
 import { ProposalActionRoleMember } from './models/proposal-action-role-member.model';
 import { ProposalActionRole } from './models/proposal-action-role.model';
 import { ProposalAction } from './models/proposal-action.model';
+import { ProposalActionsService } from './proposal-actions.service';
 import { ProposalActionEventHostsResolver } from './resolvers/proposal-action-event-hosts.resolver';
 import { ProposalActionEventsResolver } from './resolvers/proposal-action-events.resolver';
 import { ProposalActionGroupConfigsResolver } from './resolvers/proposal-action-group-configs.resolver';
 import { ProposalActionRolesResolver } from './resolvers/proposal-action-roles.resolver';
 import { ProposalActionsResolver } from './resolvers/proposal-actions.resolver';
-import { ProposalActionsService } from './proposal-actions.service';
 
 @Module({
   imports: [
